feat: 升级nginx版本，支持docker构建 (#15)
* feat：调整构建脚本，删除nginx代码

* 使用rm -f防止删除失败

* feat; 无需执行清理逻辑

* feat: 修改nginx限流的问题

* feat： 重新编写readme

* feat: 支持容器镜像的构建
