+++ .github/workflows/docker.yml
@@ -1,0 +1,41 @@
+name: DockerImage
+
+on:
+  release:
+    types: [published]
+
+jobs:
+  release:
+    name: Release Polaris Docker Image
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        goos: [linux]
+        goarch: [amd64]
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+
+      - name: Get version
+        id: get_version
+        run: echo ::set-output name=VERSION::${GITHUB_REF/refs\/tags\//}
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+        with:
+          config-inline: |
+            insecure-entitlements = [ "network.host" ]
+
+      - name: Log in to Docker Hub
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.POLARIS_DOCKER_NAME }}
+          password: ${{ secrets.POLARIS_DOCKER_PASSWORD }}
+          
+      - name: Build Server
+        id: build-server
+        env:
+          DOCKER_TAG: ${{ steps.get_version.outputs.VERSION }}
+        run: |
+          cd build
+          ./build_docker.sh ${DOCKER_TAG}
